refactor(scalars): move text, textarea, checkbox and toggle to new UI
