+++ packages/design-system/src/scalars/components/boolean-field/boolean-field.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import type { CheckboxProps } from "../fragments/checkbox-field/checkbox.js";
+import type { CheckboxProps } from "../../../ui/components/data-entry/checkbox/checkbox.js";
+import type { ToggleProps } from "../../../ui/components/data-entry/toggle/toggle.js";
 import { CheckboxField } from "../fragments/checkbox-field/index.js";
 import { ToggleField } from "../fragments/toggle-field/index.js";
-import type { ToggleProps } from "../fragments/toggle-field/toggle.js";
 export interface BooleanFieldProps
   extends Omit<CheckboxProps, "defaultValue" | "onChange" | "value">,
     Omit<ToggleProps, "onChange" | "value"> {
