+++ packages/design-system/src/scalars/components/fragments/checkbox-field/checkbox-field.tsx
@@ -1,5 +1,8 @@
+import {
+  Checkbox,
+  type CheckboxProps,
+} from "../../../../ui/components/data-entry/checkbox/checkbox.js";
 import { withFieldValidation } from "../with-field-validation/with-field-validation.js";
-import { Checkbox, type CheckboxProps } from "./checkbox.js";
 
 const CheckboxField = withFieldValidation<CheckboxProps>(Checkbox);
