+++ packages/design-system/src/scalars/components/fragments/checkbox-field/index.ts
@@ -1,2 +1,1 @@
 export { CheckboxField } from "./checkbox-field.js";
-export { Checkbox } from "./checkbox.js";
