+++ packages/design-system/src/scalars/components/fragments/toggle-field/index.ts
@@ -1,2 +1,1 @@
 export { ToggleField } from "./toggle-field.js";
-export { Toggle } from "./toggle.js";
