+++ packages/design-system/src/scalars/components/fragments/toggle-field/toggle-field.tsx
@@ -1,5 +1,8 @@
+import {
+  Toggle,
+  type ToggleProps,
+} from "../../../../ui/components/data-entry/toggle/toggle.js";
 import { withFieldValidation } from "../with-field-validation/with-field-validation.js";
-import { Toggle, type ToggleProps } from "./toggle.js";
 
 const ToggleField = withFieldValidation<ToggleProps>(Toggle);
