+++ packages/design-system/src/ui/components/data-entry/checkbox/index.ts
@@ -1,0 +1,2 @@
+export { type CheckboxValue } from "./checkbox-base.js";
+export { Checkbox, type CheckboxProps } from "./checkbox.js";
