+++ packages/design-system/src/ui/components/data-entry/index.ts
@@ -1,1 +1,9 @@
-export { Input, type InputProps } from "./input/input.js";
+export {
+  Checkbox,
+  type CheckboxProps,
+  type CheckboxValue,
+} from "./checkbox/index.js";
+export { Input, type InputProps } from "./input/index.js";
+export { TextInput, type TextInputProps } from "./text-input/index.js";
+export { Textarea, type TextareaProps } from "./textarea/index.js";
+export { Toggle, type ToggleProps } from "./toggle/index.js";
