+++ packages/design-system/src/ui/components/data-entry/text-input/index.ts
@@ -1,0 +1,2 @@
+export { TextInput, type TextInputProps } from "./text-input.js";
+export { type CommonTextProps } from "./types.js";
