+++ packages/design-system/src/ui/components/data-entry/textarea/index.ts
@@ -1,0 +1,1 @@
+export { Textarea, type TextareaProps } from "./textarea.js";
