+++ packages/design-system/src/ui/components/data-entry/toggle/index.ts
@@ -1,0 +1,1 @@
+export { Toggle, type ToggleProps } from "./toggle.js";
